==> notebooks/01_explore_scrape.ipynb <==
# Initial Web Scraping

import pandas as pd
from bs4 import BeautifulSoup
import requests
import numpy as np
import bokeh
import matplotlib.pyplot as plt
import os
import lxml.html as lh

results_per_page = 10000
current_page = 1

url = 'http://www.vgchartz.com/games/games.php?page=' + str(current_page) +\
        '&results=' + str(results_per_page) +\
'&name=&console=&keyword=&publisher=&genre=&order=Sales&ownership\
=Both&boxart=Both&banner=Both&showdeleted=&region=All&goty_year=&developer=&direction\
=DESC&showtotalsales=1&shownasales=1&showpalsales=1&showjapansales=1&showothersales=1&\
showpublisher=1&showdeveloper=1&showreleasedate=1&showlastupdate=1&showvgchartzscore=1&\
showcriticscore=1&showuserscore=1&showshipped=1&alphasort=&showmultiplat=No'

Here's a cool scraper function I found, that I've modified to do what I need to do from VGChartz.

this script was originally made by Github user nicholassherwin and can be found [here](https://gist.github.com/nicholassherwin/5a442e1cd63a31411a75dda2c7b5d2a4)

I have modified it slightly for the needs of this project

def find_console_tags(soup):
    # Console tags are stored as images, so we find the image tag and record its 'alt' value as text
    consoles = list()
    for img in soup.find_all('img'):
        if 'images/consoles'in img['src']:
            # Cut file path elements from string
            console_tag = (img['src'][17:-6])
            consoles.append(img['alt'])
    return consoles


# Find the names of games from the links
def find_names_column(table_path):
    names_list = list()
    for row in table_path.xpath('.//tr'):
        for td in row.xpath('.//td'):
            if not td.find('a') is None:
                names_list.append(td.find('a').text.strip()) 
    return names_list

# Write a function that takes in a VGChartz URL and gives us all the data in their video game database
def scrape_vgchartz_videogame_db_page(url):
    try:
        response = requests.get(url)

        ### Check the Status
        assert(response.status_code == 200)," Website not OK " # status code = 200 => OK

        #Store the contents of the website under doc
        page=response.text
        soup = BeautifulSoup(page, "lxml")
        doc = lh.fromstring(response.content)

        # Selects the table with all the data in it on HTML using xpath
        target_table_path = doc.xpath('//*[@id="generalBody"]/table')[0]

        # Find column values that won't be scraped correctly with .text option
        names_list = find_names_column(target_table_path)
        consoles = find_console_tags(soup)

        # Parse non-image and non-URL info from the data table to a pandas DataFrame
        row_dict={}
        df=pd.DataFrame()
        row_list= list()
        for counter,row in enumerate(target_table_path.xpath(".//tr")):
            if counter > 2: # To skip header rows
                row_list=[td.text for td in row.xpath(".//td")]
                row_dict[counter] = row_list

        df=pd.DataFrame.from_dict(row_dict).transpose()
        df.columns = ['position','game','blank','console','publisher','developer','vgchart_score',\
                     'critic_score','user_score','total_shipped','total_sales',\
                      'na_sales','pal_sales','japan_sales','other_sales',\
                      'release_date','last_update']

        # Correct the console and game columns using scraped values

        df=df.reset_index().drop(columns = ['index','blank'])
        df['console'] = consoles
        df['game'] = names_list
        return df
    
    except:
        return pd.DataFrame()
    # We can 'hack' the URL to display any number of results per page. I'll leave it as an argument.
def scrape_all_vg_chartz_videogame_db(results_per_page):
    df = pd.DataFrame()
    current_page = 1
    games_left = True
    while games_left:
        url = 'http://www.vgchartz.com/games/games.php?page=' + str(current_page) +\
        '&results=' + str(results_per_page) + '&name=&console=&keyword=&publisher=&genre=&order=Sales&ownership\
        =Both&boxart=Both&banner=Both&showdeleted=&region=All&goty_year=&developer=&direction\
        =DESC&showtotalsales=1&shownasales=1&showpalsales=1&showjapansales=1&showothersales=1&\
        showpublisher=1&showdeveloper=1&showreleasedate=1&showlastupdate=1&showvgchartzscore=1&\
        showcriticscore=1&showuserscore=1&showshipped=1&alphasort=&showmultiplat=No'
        new_df = scrape_vgchartz_videogame_db_page(url)
        if not new_df.empty:
            df = df.append(new_df)
            print(f'page {current_page} done')
            current_page += 1
        else:
            break
    return df
# Run the code to scrape! I did 10,000 rows per page to speed things up.
df=scrape_all_vg_chartz_videogame_db(5000)

# Compress and store for later!
df.to_csv('vgsales.csv', sep=",", encoding='utf-8', index=False)
print("Wrote scraper data")

**It worked** We have a bunch of data now, let's get started

df.sample(5)

Let's clean up a little

df = df.replace('N/A', np.nan)

df.sample(5)

df.critic_score.value_counts()

start_here = df[(df.console != 'Series') & (df.critic_score.apply(str.strip) != 'N/A')]

### Saving our progesss

Export to csv our scraping data

start_here.to_csv('../data/critic_scores.csv')

response = requests.get(url)
page = response.text
soup = BeautifulSoup(page, "lxml")

start_here[start_here.game == 'Hardwood Spades']

## More random doodles


just some messing around scraper stuff

table = soup.find(id='generalBody')

infos = table.find_all('tr')[3:-1]

len(infos)

columns = ['game', 'console', 'publisher','developer', 'vgchartz_score',
           'critic_score', 'user_score', 'total_shipped', 'release_date']

vgchartz = pd.DataFrame(columns=columns)

infos[105].find_all('td')[15]

infos[100].find_all('img')[1]['alt']

lengths = [len(game) for game in infos]


for game in infos:
    stuff = game.find_all('td')
    index = stuff[0].text.strip()
    name = stuff[2].text.strip()
    console = game.find_all('img')[1]['alt'].strip()
    publisher = stuff[4].text.strip()
    developer = stuff[5].text.strip()
    shipped = stuff[9].text.strip()
    release = stuff[15].text.strip()
    vgchartz.loc[index] = [name, console, publisher, developer, 0, 0, 0, shipped, release]

no_series = vgchartz.loc[vgchartz.console != 'Series']

no_series.release_date = pd.to_datetime(no_series.release_date, dayfirst=True, errors='coerce')

def clean_sales(n):
    n = n.replace('m', '0000').replace('.', '')
    
    if n == 'N/A':
        return np.NaN
    
    return int(n)    

no_series.total_shipped = no_series.total_shipped.apply(clean_sales)

no_series
