==> notebooks/02_wikiscraping.ipynb <==
# Wikipedia Scrape Notebook

Here we will test out grabbing sections of text from Wikipedia articles, using some hobbled together code and the wikipedia module

import pandas as pd
import numpy as np
import wikipedia

So first we'll import our csv that we exported from the vgchartz scraping notebook

These are the entries that have values for 'critic_score' and that are not themselves series (i.e. we removed the Pokemon Series entry and instead just kept the individual games)

df = pd.read_csv('../data/critic_scores.csv')
df.sample(5)

df['plots'] = ''
df.shape

**Scraping time**

So this block is going to go through every game name in our list, find the corresponding Wikipedia pages, and attatch to our DataFrame the 'plot' or 'story' section

We have to have the name exactly as it appears on the page name, so we're going to lose quite a few rows naturally...hopefully it's not too much

This code was taken mostly from [this medium article](https://medium.com/@Alexander_H/scraping-wikipedia-with-python-8000fc9c9e6c) about gathering movie plots, and it was modified slightly for our purpose

# create a list of all the names you think/know the section might be called
possibles = ['Plot','Synopsis','Plot synopsis','Plot summary', 
             'Story','Plotline','Gameplay','Summary',
            'Content','Premise']

# sometimes those names have 'Edit' on accident
possibles_edit = [i + 'Edit' for i in possibles]

#then merge those two lists together
all_possibles = possibles + possibles_edit



# now for the actual fetching!
for idx, row in df.iterrows():
    
    # load the page once and save it as a variable
    try:
        wik = wikipedia.WikipediaPage(row['game'])
    except:
        
        # Sometimes the page titles are in the form 'GameName (video game)'
        try:
            wik = wikipedia.WikipediaPage(row['game']+ ' (video game)')
        except:
            wik = 'N/A'
            print('no page found')
        
        
    # a new try, except for the plot
    try:
        # for all possible titles in all_possibles list
        for j in all_possibles:
            # if that section does exist, i.e. it doesn't return 'None'
            if wik.section(j) != None:
                #then that's what the plot is! Otherwise try the next one!
                plot = wik.section(j).replace('\n','').replace("\'","")
                df.at[idx, 'plots'] = plot
                break
            
                
    # if none of those work, or if the page didn't load from above, then plot
    # equals np.NaN
    except:
        df.at[idx, 'plots'] = 'N/A'
        print(row['game'])
        continue
        

Alright we got it!!! The output of the cell above will show us all of the titles that we didn't find a page for. Going to worry about sorting through those later

df.sample(5)

**SAVE YOUR WORK**

Let's save the raw return of the scrape and then save the results that we got entries for as seperate csv files

df.to_csv('../data/wikipedia_results.csv')
df[(df.plots != '') & (df.plots != 'N/A')].to_csv('../data/games_with_plots.csv')

games_with_plots = pd.read_csv('../data/games_with_plots.csv')
games_with_plots.shape

So after scraping, we got 4196 games with full plot descriptions...that'll do for now
