==> notebooks/07_smarter_modeling.ipynb <==
# Real Smart Model Iterating

It's time to model smarter

# Standard stuff
import pandas as pd
import string

# Preprocessing
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.pipeline import Pipeline
from nltk.corpus import stopwords, wordnet
from nltk import pos_tag
from nltk.stem import WordNetLemmatizer 

# Modeling
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.ensemble import RandomForestClassifier



df = pd.read_csv('../data/final_plots.csv').drop('Unnamed: 0', axis=1)

def bin_scores(score):
    if score > 60:
        if score > 75:
            return 2
        return 1
    return 0

df['success'] = df['Critic_Score'].apply(bin_scores)

def clean_text(text):
    
    # Removes punctuation
    words = [''.join(ch for ch in s if ch not in string.punctuation)\
             for s in text.split()]
    
    # Returns the lower-case string
    return ' '.join(words).lower()

df.plots = df.plots.apply(clean_text)

stops = stopwords.words('english') + ['game', 'player', 'gameplay']
df.head()

We need to tag our parts of speech so we can lemmatize them effectively 

from nltk.corpus import wordnet

def get_wordnet_pos(treebank_tag):
    '''
    Translate nltk POS to wordnet tags
    '''
    if treebank_tag.startswith('J'):
        return wordnet.ADJ
    elif treebank_tag.startswith('V'):
        return wordnet.VERB
    elif treebank_tag.startswith('N'):
        return wordnet.NOUN
    elif treebank_tag.startswith('R'):
        return wordnet.ADV
    else:
        return wordnet.NOUN

Continuing from the last notebook, I'd like to start by making a pipeline so we can get some cross-validation going without worrying about data leakage.

For the first pipeline we'll stick with the random forest

def big_lemmatizer(text):
    tokens = [word for word in text.split() if word not in stops]
    tags = pos_tag(tokens)
    tags = [(word[0], get_wordnet_pos(word[1])) for word in tags]
    lemmatizer = WordNetLemmatizer()
    words = [lemmatizer.lemmatize(word[0], word[1]) for word in tags]
    return ' '.join(words)
df.plots = df.plots.apply(big_lemmatizer)

df.plots.apply(clean_text)
y = df['success']
X = df['plots']
X_train, X_test, y_train, y_test = train_test_split(X, y)

tf = TfidfVectorizer(stop_words='english', max_features=50)
X_train_tf = tf.fit_transform(X_train.values)
X_test_tf = tf.transform(X_test.values)



rf = RandomForestClassifier()
rf.fit(X_train_tf, y_train.values.ravel())

print('train acc: ', rf.score(X_train_tf, y_train))
print('test acc: ', rf.score(X_test_tf, y_test))

pipe = Pipeline([
    ('tf', TfidfVectorizer(stop_words=stops)),
    ('rf', RandomForestClassifier())
])

params = {'tf__max_features': (20,50,100,500),
          'tf__ngram_range': ((1,1), (1,2)),
          'rf__max_depth': (5, 10, 20, 100, None)}
gs = GridSearchCV(estimator=pipe, param_grid=params, return_train_score=True, verbose=1)

gs.fit(X_train, y_train)

gs.best_score_

from sklearn.metrics import accuracy_score, log_loss
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC, LinearSVC, NuSVC
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
classifiers = [
    KNeighborsClassifier(3),
    SVC(kernel="rbf", C=0.025, probability=True),
    NuSVC(probability=True),
    DecisionTreeClassifier(),
    RandomForestClassifier(),
    AdaBoostClassifier(),
    GradientBoostingClassifier()
    ]

for classifier in classifiers:
    
    classifier.fit(X_train_tf, y_train.values.ravel())   
    print(classifier)
    print("model score: %.3f" % classifier.score(X_test_tf, y_test.values.ravel()))

df.plots = df.Platform.apply(str.lower) + ' ' + df.plots

df.head()
